Fix issue with redirects when toggling admin mode on
