+++ frontend/src/utils/utils.ts
@@ -1,4 +1,23 @@
-import { TAIdentity, TARequestDetail, TARole, TAStatusName } from '../api/dashboard/types';
+import { TAIdentity, TARequest, TARequestDetail, TARole, TAStatusName } from '../api/dashboard/types';
+
+export const sortAndFilterRequests = (
+  requests: TARequest[],
+  sortField: string,
+  searchTerm: string
+): TARequest[] => {
+  const sortDirection = sortField.startsWith('-') ? 'desc' : 'asc';
+  const sortFieldName = sortField.replace('-', '') as keyof TARequest;
+  const filtered = searchTerm
+    ? requests.filter((r) =>
+        `${JSON.stringify(r)}`.toLowerCase().includes(searchTerm.toLowerCase())
+      )
+    : requests;
+  return [...filtered].sort((a, b) => {
+    if (a[sortFieldName]! < b[sortFieldName]!) return sortDirection === 'asc' ? -1 : 1;
+    if (a[sortFieldName]! > b[sortFieldName]!) return sortDirection === 'asc' ? 1 : -1;
+    return 0;
+  });
+};
 
 /**
  * Validates a US telephone number (10 digits, allows common formatting)
